+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -9,6 +9,8 @@ import javax.annotation.Nullable;
 import javax.sql.DataSource;
 import nextstep.jdbc.support.Assert;
 import nextstep.jdbc.support.DataAccessUtils;
+import nextstep.jdbc.support.PreparedStatementCallback;
+import nextstep.jdbc.support.RowMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
